>  compile NEMU
Linux gpu04-NF5280M4 5.4.0-150-generic #167~18.04.1-Ubuntu SMP Wed May 24 00:51:42 UTC 2023 x86_64 x86_64 x86_64 GNU/Linux
 00:38:12 up 57 days, 14:49,  2 users,  load average: 0.44, 0.16, 0.06

+++ nemu/src/monitor/sdb/sdb.c
@@ -144,7 +144,7 @@ static int cmd_p(char *args) {
   bool flag = true;
   word_t result = expr(args, &flag);
   if (flag) {
-    printf("%u\n", result);
+    printf("%u\n", (uint32_t)result);
   } else {
     printf("Failed to evaluate expression: %s\n", args);
   }
